feat(common): Add reallocating buffers and unit tests
ReallocatingBuffer.java extends OutputStream, providing a reallocating
byte[] buffer. When writes are performed, the capacity will grow (in
multiples of the capacityIncrement parameter) to fit the buffer, until
the capacity reaches the maximumCapacity parameter. The abstraction does
not limit how many bytes can be written, but the default implementation
does have a limit as it only stores bytes in memory.

ReallocatingOutputStreamBuffer extends ReallocatingBuffer and Flushable,
which is initialized with an OutputStream along with the other capacity
parameters. The implementation allows for more than maximumCapacity
bytes to be written by flushing the buffer. If the write operation is
larger than the maximum buffer capacity, then the buffer is flushed to
the OutputStream, then the operation is passed directly to the
OutputStream implementation.